backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
websockets==13.1
python-multipart==0.0.9
requests==2.32.3
aiohttp==3.10.11
psutil==6.0.0
groq==0.13.1
openai==1.51.2
httpx==0.27.2
google-generativeai==0.8.3
deepgram-sdk==3.7.7
edge-tts==6.1.13
numpy==2.1.2
pydub==0.25.1
pyautogui==0.9.54
pygetwindow==0.0.9
pyperclip==1.8.2
mss==9.0.1
pillow==10.4.0
watchdog==5.0.3
playwright==1.48.0
yt-dlp==2025.1.26
pvporcupine==3.0.3
pycaw==20240210
comtypes==1.4.8
